app/config.py
"""Configuration — loads .env and exposes settings."""

import os
from pathlib import Path

from dotenv import load_dotenv

PROJECT_ROOT = Path(__file__).resolve().parent.parent

load_dotenv(PROJECT_ROOT / ".env")

OPENROUTER_API_KEY: str = os.getenv("OPENROUTER_API_KEY", "")
DEFAULT_MODEL: str = os.getenv("DEFAULT_MODEL", "deepseek/deepseek-chat-v3-0324")
PORT: int = int(os.getenv("PORT", "8000"))

PROMPTS_DIR = PROJECT_ROOT / "app" / "prompts"
EXAMPLES_DIR = PROJECT_ROOT / "examples"

# Models offered in the UI dropdown (id, display name)
MODELS = [
    {"id": "deepseek/deepseek-chat-v3-0324", "name": "DeepSeek V3", "provider": "DeepSeek"},
    {"id": "qwen/qwen3-235b-a22b", "name": "Qwen3 235B", "provider": "Alibaba"},
    {"id": "anthropic/claude-sonnet-4", "name": "Claude Sonnet 4", "provider": "Anthropic"},
    {"id": "openai/gpt-4o", "name": "GPT-4o", "provider": "OpenAI"},
    {"id": "google/gemini-2.5-flash-preview", "name": "Gemini 2.5 Flash", "provider": "Google"},
    {"id": "meta-llama/llama-4-maverick", "name": "Llama 4 Maverick", "provider": "Meta"},
]
