requirements.txt
fastapi>=0.115
uvicorn>=0.34
openai>=1.60
beautifulsoup4>=4.12
lxml>=5.0
python-dotenv>=1.0
